refact: extrair drag & drop para DragDropActions (actor-sheet.mjs)
